rename main to game page
